+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.17.0]
+## [0.18.0] - 2018-04-02
+
+### Changed
+
+- Use scoped packages on npm:
+  - `uiengine` -> `@uiengine/core`
+  - `uiengine-theme` -> `@uiengine/theme`
+  - `uiengine-adapter-XYZ` -> `@uiengine/adapter-XYZ`
+
+## [0.17.0] - 2018-04-02
 
 ### Added